+++ rocket_league_drl/package.xml
@@ -13,4 +13,5 @@
   <exec_depend>geometry_msgs</exec_depend>
   <exec_depend>std_msgs</exec_depend>
   <exec_depend>std_srvs</exec_depend>
+  <exec_depend>arc_tutorials</exec_depend>
 </package>
