--- rocket_league_drl/params/vac.yaml
@@ -1,2 +1,0 @@
-gamma: 0.95
-learning_rate: 0.01
