--- rocket_league_drl/src/rocket_league_drl/agents/__init__.py
@@ -1,8 +1,0 @@
-from ._agent_interface import AgentInterface
-from .dqn_agent import DQNAgent
-from .vac_agent import VACAgent
-
-__all__ = [
-    "AgentInterface",
-    "DQNAgent",
-    "VACAgent"]
