+++ rocket_league_drl/src/rocket_league_drl/cartpole_direct_interface.py
@@ -1,64 +1,63 @@
 #!/usr/bin/env python
 
 # package
-from rocket_league_drl.interfaces import ROSInterface
+from rocket_league_drl import ROSInterface
 
 # ROS
 import rospy
 
 # System
 import gym
-from enum import IntEnum, unique, auto
-from threading import Condition
 
 class CartpoleDirectInterface(ROSInterface):
     """ROS interface for the cartpole game."""
+    _node_name = "cartpole_direct"
     def __init__(self):
-        rospy.init_node('cartpole_direct_drl')
-        self._cond = Condition()
+        super().__init__()
+
         self._RENDER = rospy.get_param('~render', False)
 
         self._env = gym.make('CartPole-v0')
-        self._obs = self._env.reset() 
+        self._obs = None
         self._reward = 0
         self._done = 0
         self._info = {}
 
     @property
-    def OBSERVATION_SIZE(self):
-        """The observation size for the network."""
-        return self._env.observation_space.shape[0]
+    def action_space(self):
+        """The Space object corresponding to valid actions."""
+        return self._env.action_space
 
     @property
-    def ACTION_SIZE(self):
-        """The action size for the network."""
-        return self._env.action_space.n
+    def observation_space(self):
+        """The Space object corresponding to valid observations."""
+        return self._env.observation_space
 
-    def reset_env(self):
+    def _reset_env(self):
         """Reset environment for a new training episode."""
         self._obs = self._env.reset() 
         self._reward = 0
         self._done = 0
         self._info = {}
 
-    def reset(self):
+    def _reset_self(self):
         """Reset internally for a new episode."""
         pass
 
-    def has_state(self):
+    def _has_state(self):
         """Determine if the new state is ready."""
         return True
 
-    def clear_state(self):
+    def _clear_state(self):
         """Clear state variables / flags in preparation for new ones."""
         pass
 
-    def get_state(self):
+    def _get_state(self):
         """Get state tuple (observation, reward, done, info)."""
         assert self.has_state()
         return (self._obs, self._reward, self._done, self._info)
 
-    def publish_action(self, action):
+    def _publish_action(self, action):
         """Publish an action to the ROS network."""
         assert action >= 0 and action < self.ACTION_SIZE
         if self._RENDER:
