--- rocket_league_drl/src/rocket_league_drl/interfaces/__init__.py
@@ -1,10 +1,0 @@
-from ._ros_interface import ROSInterface
-from .cartpole_interface import CartpoleInterface
-from .cartpole_direct_interface import CartpoleDirectInterface
-from .snake_interface import SnakeInterface
-
-__all__ = [
-    "ROSInterface",
-    "CartpoleInterface",
-    "CartpoleDirectInterface",
-    "SnakeInterface"]
