--- rocket_league_drl/src/rocket_league_drl/make_evaluator.py
@@ -1,65 +1,0 @@
-"""
-Contains the make_evaluator factory.
-
-License:
-  BSD 3-Clause License
-  Copyright (c) 2021, Autonomous Robotics Club of Purdue (Purdue ARC)
-  All rights reserved.
-  Redistribution and use in source and binary forms, with or without
-  modification, are permitted provided that the following conditions are met:
-  1. Redistributions of source code must retain the above copyright notice, this
-      list of conditions and the following disclaimer.
-  2. Redistributions in binary form must reproduce the above copyright notice,
-      this list of conditions and the following disclaimer in the documentation
-      and/or other materials provided with the distribution.
-  3. Neither the name of the copyright holder nor the names of its
-      contributors may be used to endorse or promote products derived from
-      this software without specific prior written permission.
-  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-  AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-  IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-  DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-  FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-  DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-  OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-"""
-
-from rocket_league_drl.interfaces import ROSInterface
-
-import rospy
-from std_msgs.msg import String
-
-def makeEvaluator(cls):
-    """Create an evaluator extending the provided agent."""
-    assert issubclass(cls, ROSInterface)
-
-    class Evaluator(cls):
-        """Class capable of evaluating an agent against live input."""
-        def __init__(self):
-            super().__init__()
-
-            # initialize variables
-            self.reset_env()
-            self.reset()
-
-            while not rospy.is_shutdown():
-                try:
-                    while not self.wait_for_state():
-                        pass
-                except rospy.ROSInterruptException:
-                    pass
-
-                state = self.get_state()
-
-                # check for end of game and react
-                if not state[2]:
-                    self.clear_state()
-                    self.publish_action(self.agent.eval(state))
-                else:
-                    self.reset_env()
-                    self.reset()
-
-    return Evaluator
